EERTREE.Node

Documentation

>>> :set -XTypeApplications -XDataKinds

data Node (n :: Nat)

A node corresponsing to a palindrome in an alphabet of size n.

Constructors

Node

Fields

index :: !Integer

len :: !Int

value :: Seq (Symbol n)

parent :: Maybe (Symbol n, Node n)

ancestors :: Vector (Node n)

edges :: Vector (Weakly (Node n))

links :: IntMap (Node n)

symbol :: Node n -> Maybe (Symbol n)

First (last) symbol of a palindrome.

>>> symbol (fromPalindrome @2 [1,0,1])
Just 1
>>> symbol (fromPalindrome @2 [])
Nothing

edge :: Symbol n -> Node n -> Node n

Follow an edge in palindromic tree. This corresponds to adding a symbol to both ends of some palindrome.

>>> edge 1 (fromPalindrome @2 [0,1,0])
fromPalindrome [1,0,1,0,1]
>>> edge 0 (fromPalindrome @2 [])
fromPalindrome [0,0]

directLink :: KnownNat n => Symbol n -> Node n -> Node n

Follow a direct link to find the largest palindrome suffix preceded by a given symbol in another palindrome.

>>> directLink 1 (fromPalindrome @2 [0,1,0,1,0])
fromPalindrome [0,1,0]
>>> directLink 0 (fromPalindrome @2 [0,1,0,1,0])
fromPalindrome []
>>> directLink 1 (fromPalindrome @2 [0,0,0,0,0])
oddNode

link :: KnownNat n => Node n -> Node n

The largest palindrome suffix

>>> link (fromPalindrome @2 [1,0,0,0,1])
fromPalindrome [1]

>>> link (fromPalindrome @2 [0,0,0])
fromPalindrome [0,0]

fromPalindrome :: KnownNat n => [Symbol n] -> Node n

Construct a Node n from a palindrome.

NOTE: this will result in an error if input is not a palindrome.

>>> fromPalindrome @2 [0,1,1]
*** Exception: not a palindrome
...

symbolAt :: Int -> Node n -> Maybe (Symbol n)

Like symbolAt', but tries to minimize jumps.

>>> symbolAt 3 (fromPalindrome @2 [0,1,0,1,0,1,0])
Just 1
>>> pal = fromPalindrome @2 [0,1,0,1,0,1,0]
>>> symbolAt 1 pal == symbolAt 5 pal
True
>>> pal = fromPalindrome @2 [0,1,0,1,0,1,0]
>>> symbolAt 3 pal == symbolAt' 3 pal
True

symbolAt' :: Int -> Node n -> Maybe (Symbol n)

Find a symbol at a given position in a palindrome. Note that it does not matter from what end you start indexing.

>>> symbolAt' 0 (fromPalindrome @2 [0,1,0])
Just 0
>>> symbolAt' 1 ((fromPalindrome @2 [0,1,0]))
Just 1
>>> symbolAt' 2 ((fromPalindrome @2 [0,1,0]))
Nothing
>>> symbolAt' 3 (fromPalindrome @2 [0,1,0])
Nothing
>>> symbolAt' (-1) (fromPalindrome @2 [0,1,0])
Nothing

pathTo :: Node n -> [Symbol n]

Path to the node from one of the roots. This corresponds to the left half of the palindrome a node represents.

>>> pathTo (fromPalindrome @2 [1,1,0,1,1])
[1,1,0]

getAncestors :: Node a -> Vector (Node a)

Build a Vector of a node's parent ancestors that are $2^i$ edges away from the node.

>>> (F.toList . value) <$> getAncestors (fromPalindrome @2 [1,0,1,0,1,0,1])
[[1,0,1,0,1,0,1],[0,1,0,1,0],[0]]

mkEdge :: forall n. KnownNat n => Symbol n -> Node n -> Node n

Construct a new node by following an edge. This corresponds to adding a symbol to both ends of a palindrome.

>>> mkEdge 0 (fromPalindrome @2 [])
fromPalindrome [0,0]
>>> mkEdge 1 (fromPalindrome @2 [0])
fromPalindrome [1,0,1]

newSuffixOf :: KnownNat n => Symbol n -> Node n -> Node n

Find what would be the largest suffix of a new palindrome after added a given symbol to the end of another palindrome.

>>> newSuffixOf 1 (fromPalindrome @2 [1,0,1,0,1])
fromPalindrome [1,1]
>>> newSuffixOf 0 (fromPalindrome @2 [1,0,1,0,1])
fromPalindrome [0,1,0,1,0]
>>> newSuffixOf 1 (fromPalindrome @2 [0,0,0,0,0])
fromPalindrome [1]

mkDirectLinks

Arguments

:: KnownNat n

=> Symbol n
Symbol to add.

-> Node n
Parent palindrome.

-> IntMap (Node n)
Direct links for child palindrome.

Construct direct links for a child node, given its parent and a symbol to add to both ends.

>>> mkDirectLinks 0 (fromPalindrome @2 [0,1,0,1,0])
fromList [(0,fromPalindrome [0]),(1,fromPalindrome [0,0])]
>>> mkDirectLinks 1 (fromPalindrome @2 [0,1,0,1,0])
fromList [(0,fromPalindrome [1,0,1,0,1]),(1,fromPalindrome [])]

evenNode :: forall n. KnownNat n => Node n

An even node corresponding to an empty palindrome.

>>> evenNode @2
fromPalindrome []
>>> edge 1 (evenNode @2)
fromPalindrome [1,1]

oddNode :: forall n. KnownNat n => Node n

An odd node, corresponding to a parent of singleton palindromes.

>>> edge 1 (oddNode @2)
fromPalindrome [1]

Note that there is no palindrome that corresponds to this node:

>>> oddNode @2
oddNode