EERTREE.Node.Internal.Weakly

Documentation

data Weakly a

Weakly stores a weak link to the result of some function application. Weak link means that the result can be garbage collected after its been used even if weak links still exist.

However, Weakly a also contains a recipe of how to constuct that value if it is needed again in the future.

This makes Weakly useful when working with large lazy graph-like data structures to reduce memory usage when a working part of that structure changes.

Constructors

forall b. Weakly

Fields

wrRecipe :: b -> a

How to (re)construct a value.

wrIngredient :: b

Source.

wrResult :: IORef (Weak a)

A weak link to result.

applyWeakly :: (a -> b) -> a -> Weakly b

Apply function to an argument producing a result that is weakly referenced (and hence can be garbage collected).

fromWeakly :: Weakly b -> b

Extract weakly referenced result, reevaluating if necessary.