EERTREE.Random

Documentation

>>> :set -XTypeApplications -XDataKinds -XOverloadedStrings

randomSymbolsIO :: KnownNat n => Int -> IO [Symbol n]

Generate list of random symbols of length len

genRandomSymbols @4 10

3,2,3,1,2,3,0,0,2,0

randomSymbols :: KnownNat n => Int -> Gen [Symbol n]

Generator for list of random symbols of length len

randomEERTREE :: KnownNat n => Int -> Gen (EERTREE n)

Generator for eertree from random list of symbols of length len

randomEERTREEpairIO :: KnownNat n => Int -> Int -> IO (EERTREE n, EERTREE n)

Generate pair of random eertrees of lengths len1 and len2

randomMerge :: forall n. KnownNat n => Proxy n -> Int -> Gen Int

Count number of new palindromes after merging two random eertrees

genAverageMerge :: forall n. KnownNat n => Proxy n -> Int -> Int -> Gen Double

> generate (genAverageMerge (Proxy 2) 100 500)
7.416
> generate (genAverageMerge (Proxy 4) 100 500)
3.09
> generate (genAverageMerge (Proxy @33) 100 500)
2.024