EERTREE.Simple

Documentation

>>> :set -XTypeApplications -XDataKinds -XOverloadedStrings

data EERTREE n

A palindromic tree for some string with auxillary information.

Constructors

EERTREE

Fields

strLen :: !Int

Length of the analysed string.

maxPrefix :: Node n

Maximum palindromic prefix.

maxSuffix :: Node n

Maximum palindromic suffix

strReversedPrefix :: Seq (Symbol n)

Prefix, preceding maximum palindromic suffix

strSuffix :: Seq (Symbol n)

Suffix, following maximum palindromic prefix.

palindromes :: [Node n]

Accumulated list of encountered palindromes.

empty :: forall n. KnownNat n => EERTREE n

An empty eertree.

singleton :: KnownNat n => Symbol n -> EERTREE n

An eertree for a singleton string.

>>> fromEERTREE (singleton @2 0)
[0]

eertree :: KnownNat n => [Symbol n] -> EERTREE n

Analyse a string by building an eertree.

>>> fromEERTREE (eertree @2 [0,1,0,0,1])
[0,1,0,0,1]

eertreeFromString :: KnownNat n => String -> EERTREE n

Build an eertree from string

>>> eertreeFromString @2 "01001" == eertree @2 [0,1,0,0,1]
True

reverseEERTREE :: KnownNat n => EERTREE n -> EERTREE n

EERTREE of a reversed string

> reverseEERTREE @2 "01001" == [1,0,0,1,0]

True

fromEERTREE :: EERTREE n -> [Symbol n]

Get the string back from an eertree.

>>> fromEERTREE @2 "01001"
[0,1,0,0,1]

eertreeToSeq :: EERTREE n -> Seq (Symbol n)

reverseFromEERTREE :: EERTREE n -> [Symbol n]

Get the reversed string from an eertree

>>> reverseFromEERTREE @2 "01001"
[1,0,0,1,0]

eertreeToSeqReversed :: EERTREE n -> Seq (Symbol n)

prepend :: KnownNat n => Symbol n -> EERTREE n -> EERTREE n

Add a symbol to the beginning of a string corresponding to an eertree.

>>> fromEERTREE (prepend 0 (eertreeFromString @2 "01001"))
[0,0,1,0,0,1]

append :: KnownNat n => Symbol n -> EERTREE n -> EERTREE n

Add a symbol to the end of a string corresponding to an eertree

>>> fromEERTREE (append 0 (eertreeFromString @2 "01001"))
[0,1,0,0,1,0]

merge :: KnownNat n => EERTREE n -> EERTREE n -> EERTREE n

Merge two eertrees in O(1) on average

>>> fromEERTREE (merge @2 "0110100" "11001001")
[0,1,1,0,1,0,0,1,1,0,0,1,0,0,1]

>>> fromEERTREE (merge @2 "10010011" "0010110")
[1,0,0,1,0,0,1,1,0,0,1,0,1,1,0]

mergeLinear :: KnownNat n => EERTREE n -> EERTREE n -> EERTREE n

Merge two eertrees in min(|S1|, |S2|)

>>> fromEERTREE (mergeLinear @2 "0110100" "11001001")
[0,1,1,0,1,0,0,1,1,0,0,1,0,0,1]

>>> fromEERTREE (mergeLinear @2 "10010011" "0010110")
[1,0,0,1,0,0,1,1,0,0,1,0,1,1,0]

Applications

subpalindromes :: KnownNat n => [Symbol n] -> [[Symbol n]]

Unique subpalindromes of a string.

>>> subpalindromes @2 [0,1,0,0,1]
[[0,1,0],[1,0,0,1],[0,0],[0],[1]]

a216264 :: Int -> [Int]

Compute first \(n\) elements of A216264 sequence (binary rich strings count for \(n = 0, 1, \ldots\)).

For memory efficiency the whole sequence segment is computed at once and not lazily as one might expect.

This should run in \(\mathcal{O}(n)\) memory with garbage collector working normally.

>>> a216264 15
[1,2,4,8,16,32,64,128,252,488,932,1756,3246,5916,10618]

palindromicRefrain :: KnownNat n => EERTREE n -> (Node n, Int)

Palindromic refrain: for a given string S find a palindrome P maximizing the value |P| * occ(S, P), where occ(S, P) is the number of occurences of P in S

>>> palindromicRefrain (eertreeFromString @3 "0102010")
(fromPalindrome [0,1,0,2,0,1,0],7)

>>> palindromicRefrain (eertreeFromString @1 "000")
(fromPalindrome [0,0],4)

frequency :: KnownNat n => EERTREE n -> [(Node n, Int)]

Compute the number of occurences for each subpalindrome

>>> frequency (eertreeFromString @2 "10101")
[(fromPalindrome [1,0,1,0,1],1),(fromPalindrome [0,1,0],1),(fromPalindrome [1,0,1],2),(fromPalindrome [1],3),(fromPalindrome [0],2)]

frequency' :: KnownNat n => EERTREE n -> Map Integer (Node n, Int)

Compute the number of occurences for each subpalindrome. Return Map of frequences with Node index as a key.

>>> frequency' (eertreeFromString @2 "1000")
fromList [(-4,(fromPalindrome [0,0,0],1)),(-3,(fromPalindrome [1],1)),(-2,(fromPalindrome [0],3)),(2,(fromPalindrome [0,0],2))]

Helpers

dfsCountLevels

Arguments

:: Int
Number of levels to cover (max depth).

-> a
Root.

-> (a -> [a])
How to get subtrees.

-> [Int]
Number of elements at every level.

Efficiently count nodes at every level of a tree using DFS.

Memory efficiency comes from the fact that in depth-first traversal we don't have to store all nodes at a level and only need to store a path from root to the current node.

richSubEERTREEs :: forall n. KnownNat n => EERTREE n -> [EERTREE n]

Efficiently compute all rich eertrees that can be produced from a given one by prepending a symbol to it.

Efficiency comes from the fact that we only need to check that a new maximum prefix is a new palindrome.